backend/app/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
urllib3==1.25.10
